// ExposeInputEvents/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    Controller: ComponentFramework.PropertyTypes.StringProperty;
    onBlurScript: ComponentFramework.PropertyTypes.StringProperty;
    onChangeScript: ComponentFramework.PropertyTypes.StringProperty;
    onCopyScript: ComponentFramework.PropertyTypes.StringProperty;
    onCutScript: ComponentFramework.PropertyTypes.StringProperty;
    onClickScript: ComponentFramework.PropertyTypes.StringProperty;
    ondblClickScript: ComponentFramework.PropertyTypes.StringProperty;
    onFocusScript: ComponentFramework.PropertyTypes.StringProperty;
    onFocusInScript: ComponentFramework.PropertyTypes.StringProperty;
    onFocusOutScript: ComponentFramework.PropertyTypes.StringProperty;
    onInputScript: ComponentFramework.PropertyTypes.StringProperty;
    onKeyDownScript: ComponentFramework.PropertyTypes.StringProperty;
    onKeyUpScript: ComponentFramework.PropertyTypes.StringProperty;
    onKeyPressScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseDownScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseEnterScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseLeaveScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseMoveScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseOverScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseOutScript: ComponentFramework.PropertyTypes.StringProperty;
    onMouseUpScript: ComponentFramework.PropertyTypes.StringProperty;
    onPasteScript: ComponentFramework.PropertyTypes.StringProperty;
    onSelectScript: ComponentFramework.PropertyTypes.StringProperty;
    onWheelScript: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    Controller?: string;
}